fix: 질문 조회수 버그 수정했습니다.

+++ src/main/java/com/coverflow/question/domain/Question.java
@@ -94,7 +94,7 @@ public class Question extends BaseTimeEntity {
         this.questionStatus = request.questionStatus();
     }
 
-    public void updateViewCount(int viewCount) {
+    public void updateViewCount(long viewCount) {
         this.viewCount = viewCount;
     }
